Refactor: employeewage program to caluculate monthlywage for multiple companies
